+++ src/test/java/com/imsweb/algorithms/prcdauiho/PrcdaUihoUtilsTest.java
@@ -16,8 +16,6 @@ import static com.imsweb.algorithms.prcdauiho.PrcdaUihoUtils.MIXED_PRCDA;
 import static com.imsweb.algorithms.prcdauiho.PrcdaUihoUtils.PRCDA_NO;
 import static com.imsweb.algorithms.prcdauiho.PrcdaUihoUtils.PRCDA_UNKNOWN;
 import static com.imsweb.algorithms.prcdauiho.PrcdaUihoUtils.PRCDA_YES;
-import static com.imsweb.algorithms.prcdauiho.PrcdaUihoUtils.UIHO_FACILITY_NONE;
-import static com.imsweb.algorithms.prcdauiho.PrcdaUihoUtils.UIHO_FACILITY_UNKNOWN;
 import static com.imsweb.algorithms.prcdauiho.PrcdaUihoUtils.UIHO_NO;
 import static com.imsweb.algorithms.prcdauiho.PrcdaUihoUtils.UIHO_UNKNOWN;
 
@@ -33,15 +31,14 @@ public class PrcdaUihoUtilsTest {
     @Test
     public void testComputePrcdaUiho() {
 
-        List<String> validStates = Arrays.asList("AK", "AZ", "CO", "GA", "MN", "NV", "SE", "VA");
+        List<String> validStates = Arrays.asList("AL", "AK", "AZ", "CO", "GA", "MN", "NV", "SE", "VA");
         List<String> invalStates = Arrays.asList("", "99", null);
         List<String> validCounty = Arrays.asList("001", "003", "005", "007", "017", "097", "999");
         List<String> invalCounty = Arrays.asList("", "ABC", "01", null);
 
-        List<String> prcdaYes = Arrays.asList("AZ001", "AZ003", "AZ005", "AZ007", "AZ017", "CO007", "MN001", "MN005", "MN007", "MN017", "NV001", "NV003", "NV005", "NV007", "NV017", "SE007", "SE017",
-                "VA097");
-        List<String> prcdaNo = Arrays.asList("AZ097", "CO001", "CO003", "CO005", "CO017", "CO097", "MN003", "MN097", "NV097", "SE001", "SE003", "SE005", "SE097", "VA001", "VA003", "VA005", "VA007",
-                "VA017");
+        List<String> prcdaYes = Arrays.asList("AL003", "AZ001", "AZ003", "AZ005", "AZ007", "AZ017", "CO007", "MN001", "MN005", "MN007", "MN017", "MN097", "NV001", "NV003", "NV005", "NV007", "NV017", "SE007", "SE017",
+                "VA003", "VA097");
+        List<String> prcdaNo = Arrays.asList("AZ097", "CO001", "CO003", "CO005", "CO017", "CO097", "MN003", "NV097", "SE001", "SE003", "SE005", "SE097", "VA001", "VA005", "VA007", "VA017");
         List<String> uihoNo = Arrays.asList("AK001", "AK003", "AK005", "AK007", "AK017", "AK097", "AZ001", "AZ003", "AZ007", "AZ017", "AZ097", "CO003", "CO007", "CO017", "CO097", "GA001", "GA003",
                 "GA005", "GA007", "GA017", "GA097", "MN001", "MN003", "MN005", "MN007", "MN097", "NV003", "NV007", "NV017", "NV097", "SE001", "SE003", "SE005", "SE007", "SE017", "SE097", "VA001",
                 "VA003", "VA005", "VA007", "VA017", "VA097");
@@ -67,7 +64,6 @@ public class PrcdaUihoUtilsTest {
                 if (invalStates.contains(input.getAddressAtDxState())) {
                     Assert.assertEquals(PRCDA_UNKNOWN, output.getPRCDA());
                     Assert.assertEquals(UIHO_UNKNOWN, output.getUIHO());
-                    Assert.assertEquals(UIHO_FACILITY_UNKNOWN, output.getUIHOFacility());
                 }
                 else {
                     if (ENTIRE_STATE_PRCDA.contains(input.getAddressAtDxState())) {
@@ -94,32 +90,20 @@ public class PrcdaUihoUtilsTest {
                     }
 
                     // UIHO, UIHO FACILITY
-                    if (invalCounty.contains(input.getAddressAtDxCounty())) {
+                    if (invalCounty.contains(input.getAddressAtDxCounty()))
                         Assert.assertEquals(UIHO_UNKNOWN, output.getUIHO());
-                        Assert.assertEquals(UIHO_FACILITY_UNKNOWN, output.getUIHOFacility());
-                    }
                     else {
                         String stCnty = state + county;
-                        if ("999".equals(county) || uihoNo.contains(stCnty)) {
+                        if ("999".equals(county) || uihoNo.contains(stCnty))
                             Assert.assertEquals(UIHO_NO, output.getUIHO());
-                            Assert.assertEquals(UIHO_FACILITY_NONE, output.getUIHOFacility());
-                        }
-                        else if ("AZ005".equals(stCnty)) {
+                        else if ("AZ005".equals(stCnty))
                             Assert.assertEquals("1", output.getUIHO());
-                            Assert.assertEquals("01", output.getUIHOFacility());
-                        }
-                        else if ("CO001".equals(stCnty) || "CO005".equals(stCnty)) {
+                        else if ("CO001".equals(stCnty) || "CO005".equals(stCnty))
                             Assert.assertEquals("1", output.getUIHO());
-                            Assert.assertEquals("12", output.getUIHOFacility());
-                        }
-                        else if ("MN017".equals(stCnty)) {
+                        else if ("MN017".equals(stCnty))
                             Assert.assertEquals("1", output.getUIHO());
-                            Assert.assertEquals("49", output.getUIHOFacility());
-                        }
-                        else if ("NV001".equals(stCnty) || "NV005".equals(stCnty)) {
+                        else if ("NV001".equals(stCnty) || "NV005".equals(stCnty))
                             Assert.assertEquals("1", output.getUIHO());
-                            Assert.assertEquals("24", output.getUIHOFacility());
-                        }
                     }
                 }
             }